refactor: Extract HaveMomentoErrorCode custom matcher
It could be used in any test where it was, but now its in a shared location.
